requirements.txt
bcrypt==3.2.0
bleach==3.2.1
cffi==1.14.3
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
pycparser==2.20
PyMySQL==0.10.1
pyparsing==2.4.7
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.3.20
webencodings==0.5.1
Werkzeug==1.0.1
